# [SPARK-24410][SQL][Core] Optimization for Union outputPartitioning