--- a/sql/core/src/main/scala/org/apache/spark/sql/execution/basicPhysicalOperators.scala
+++ b/sql/core/src/main/scala/org/apache/spark/sql/execution/basicPhysicalOperators.scala
@@ -19,6 +19,7 @@ package org.apache.spark.sql.execution
 
 import scala.concurrent.{ExecutionContext, Future}
 import scala.concurrent.duration.Duration
+import scala.util.Try
 
 import org.apache.spark.{InterruptibleIterator, Partition, SparkContext, TaskContext}
 import org.apache.spark.rdd.{EmptyRDD, PartitionwiseSampledRDD, RDD}
@@ -563,12 +564,71 @@ case class RangeExec(range: org.apache.spark.sql.catalyst.plans.logical.Range)
  * [[org.apache.spark.sql.catalyst.plans.logical.Union.maxRowsPerPartition]].
  */
 case class UnionExec(children: Seq[SparkPlan]) extends SparkPlan {
+
+  /**
+   * SPARK-24410: We can't use `sameResult` or `semanticEquals` to test if we can use children's
+   * `outputPartitioning`. We can use first child's `outputPartitioning` only if partitioning
+   * expressions have the same expressions and bounding same ordinal references.
+   */
+  override def outputPartitioning: Partitioning = Try {
+    val reducedPartitioning: Partitioning = children.map { child =>
+      // Bounding partition expressions to children's outputs.
+      child.outputPartitioning match {
+        case HashPartitioning(expr, num) =>
+          val bound = expr.map(BindReferences.bindReference(_,
+            child.output.map(_.withNullability(true))))
+          HashPartitioning(bound, num)
+        case RangePartitioning(ordering, num) =>
+          val bound = ordering.map(BindReferences.bindReference(_,
+            child.output.map(_.withNullability(true))))
+          RangePartitioning(bound, num)
+        case o => o
+      }
+    }.reduceLeft((left: Partitioning, right: Partitioning) => (left, right) match {
+      case (SinglePartition, SinglePartition) => SinglePartition
+      case (HashPartitioning(exprs1, p1), HashPartitioning(exprs2, p2)) if p1 == p2 =>
+        val matched = exprs1.length == exprs2.length && exprs1.zip(exprs2).forall {
+          case (l, r) => l.semanticEquals(r)
+        }
+
+        if (matched) {
+          HashPartitioning(exprs1, p1)
+        } else {
+          super.outputPartitioning
+        }
+      case (RangePartitioning(ordering1, p1), RangePartitioning(ordering2, p2)) if p1 == p2 =>
+        val minSize = Seq(ordering1.size, ordering2.size).min
+        if (ordering1.take(minSize) == ordering2.take(minSize)) {
+          RangePartitioning(ordering1.take(minSize), p1)
+        } else {
+          super.outputPartitioning
+        }
+      case _ => super.outputPartitioning
+    })
+
+    reducedPartitioning match {
+      case HashPartitioning(_, _) => children.head.outputPartitioning
+      case RangePartitioning(ordering, num) =>
+        val newOrdering = children.head.outputPartitioning
+          .asInstanceOf[RangePartitioning].ordering.take(ordering.size)
+        RangePartitioning(newOrdering, num)
+      case _ => reducedPartitioning
+    }
+    // We can hit exception when bounding partitioning expressions. In that cases, simply
+    // use default `outputPartitioning`.
+  }.getOrElse(super.outputPartitioning)
+
   override def output: Seq[Attribute] =
     children.map(_.output).transpose.map(attrs =>
       attrs.head.withNullability(attrs.exists(_.nullable)))
 
-  protected override def doExecute(): RDD[InternalRow] =
-    sparkContext.union(children.map(_.execute()))
+  protected override def doExecute(): RDD[InternalRow] = outputPartitioning match {
+    case UnknownPartitioning(_) => sparkContext.union(children.map(_.execute()))
+    case _ =>
+      // If we have known output partitioning, it means there is the same partitioning among
+      // union children. Simply zipping RDDs without shuffling.
+      children.head.execute().zipRDDs(children.tail.map(_.execute()), true)
+  }
 }
 
 /**