## What changes were proposed in this pull request?

Currently `Union` has only unknown output partitioning. That said if the children are well partitioned (e.g., bucketed tables), we still run shuffling on the data on later operators such as aggregation.

This patch tries to better decide output partitioning for `Union` operator. So it can eliminate unnecessary shuffling, e.g.,

```scala
val N = 10

spark.range(N).selectExpr("id as key", "id % 2 as t1", "id % 3 as t2").repartition(col("key")).write.mode("overwrite").bucketBy(3, "key").sortBy("t1").saveAsTable("a1")
spark.range(N).selectExpr("id as key", "id % 2 as t1", "id % 3 as t2").repartition(col("key")).write.mode("overwrite").bucketBy(3, "key").sortBy("t1").saveAsTable("a2")
spark.conf.set("spark.sql.autoBroadcastJoinThreshold", -1)

val df = sql("select key,count(*) from (select * from a1 union all select * from a2)z group by key")
df.explain
```

```
== Physical Plan ==
*(3) HashAggregate(keys=[key#36L], functions=[count(1)])
+- *(3) HashAggregate(keys=[key#36L], functions=[partial_count(1)])
   +- Union
      :- *(1) Project [key#36L]
      :  +- *(1) FileScan parquet default.a1[key#36L] Batched: true, Format: Parquet, Location: InMemoryFileIndex[file:/root/repos/spark-1/spark-warehouse/a1], PartitionFilters: [], PushedFilters: [], ReadSchema: struct<key:bigint>
      +- *(2) Project [key#39L]
         +- *(2) FileScan parquet default.a2[key#39L] Batched: true, Format: Parquet, Location: InMemoryFileIndex[file:/root/repos/spark-1/spark-warehouse/a2], PartitionFilters: [], PushedFilters: [], ReadSchema: struct<key:bigint>
```


This patch adds a private API `zipRDDs` to `RDD`. Since `zipPartitions` asks a function to run on elements of rdds, it only supports fixed number of rdds. But for `Union`, the number of children is not fixed.

This optimization can be controlled by a SQL config `spark.sql.unionInSamePartition` which is false by default.

## How was this patch tested?

Added tests.