@viirya 2018-06-05
cc @cloud-fan I don't add tests for now yet, but would like to see if this approach looks good for you first. Thanks.