@viirya 2018-06-06
Tests are added. cc @kiszk @mgaido91