@viirya 2018-06-06
@cloud-fan This is removed of WIP and can be review now. Please take a look when you are available, as supposed that you'll be busy in this week. Thanks.