@viirya 2018-06-06
When the condition is satisfied and we know children of Union have same partitioning, this goes to let the first partition of union result includes first partitions of children RDDs, and 2nd, 3rd partitition...

I'm not sure what part the data transfer might happen?