@mgaido91 2018-06-07
Thanks for you benchmark @viirya. The performance improvement is sensible. And seems no performance regression in the other case. Can we have a similar benchmark also with records with more complex schema? Thanks.