@viirya 2018-06-12
@mgaido91 WDYT? Does the benchmark make sense to you?