@viirya 2018-07-16
I'd like to close this for now. Wait for necessary change on statistics.